Move config entry migrations to migration module
